requirements.txt
openstacksdk
python-ldap
